Relacionamento base e visão do perfil de outros usuários

+++ app/Http/Controllers/ImovelController.php
@@ -2,23 +2,28 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use App\Models\Imovel;
+use Illuminate\Http\Response;
+use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 
 class ImovelController extends Controller
-{   
+{
 
- 
+
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|Response
      */
     public function index(Request $request)
     {
-       
+
 
         $cidade = $request['cidade'];
         $tipo = $request['tipo'];
@@ -26,7 +31,7 @@ class ImovelController extends Controller
         if($cidade){
             $imoveis = DB::table('imoveis')->where('cidadeEndereco','=', $cidade)->paginate(5);
         }elseif($tipo){
-            $imoveis = DB::table('imoveis')->where('tipo', '=', $tipo)->paginate(5); 
+            $imoveis = DB::table('imoveis')->where('tipo', '=', $tipo)->paginate(5);
         }
 
         else{
@@ -42,23 +47,23 @@ class ImovelController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|Response
      */
     public function create()
-    {   
-     
+    {
+
         return view('imoveis.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Application|RedirectResponse|Response|Redirector
      */
     public function store(Request $request)
-    {   
-        $validate = $request->validate([
+    {
+        $request->validate([
             "descricao" => "required",
             "logradouroEndereco"=> "required",
             "bairroEndereco" => "required",
@@ -71,7 +76,7 @@ class ImovelController extends Controller
             "finalidade" => "required"
         ]);
 
-       
+
        $data = $request->all();
         $data['user_id'] = Auth::user()->id;
        Imovel::create($data);
@@ -82,15 +87,15 @@ class ImovelController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|Response
      */
     public function show($id)
     {
         $imovel = Imovel::find($id);
         if($imovel === null){
             $imovel = Imovel::latest()->first();
-                        
+
         }
 
         return view('imoveis.show', compact('imovel'));
@@ -100,10 +105,10 @@ class ImovelController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Application|RedirectResponse|Response|Redirector
      */
     public function edit($id)
-    {   
+    {
         $imovel = Imovel::find($id);
         if($imovel->user->id !== Auth::user()->id){
             return redirect('/')->withErrors(['msg'=>'Você só pode alterar os seus próprios imóveis']);
@@ -115,13 +120,13 @@ class ImovelController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param Request $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Application|RedirectResponse|Response|Redirector
      */
     public function update(Request $request, $id)
     {
-        $validate = $request->validate([
+             $request->validate([
             "descricao" => "required",
             "logradouroEndereco"=> "required",
             "bairroEndereco" => "required",
@@ -146,17 +151,25 @@ class ImovelController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Application|RedirectResponse|Response|Redirector
      */
     public function destroy($id)
-    {
-        Imovel::find($id)->delete();
+        {
+            if(Imovel::find($id)->user->id !== Auth::user()->id){
+                return redirect('/')->withErrors(['msg'=>'Você só pode alterar os seus próprios imóveis']);
+            }
+
+            Imovel::find($id)->delete();
+
         return redirect('/');
     }
 
 
     public function remover($id){
         $imovel = Imovel::find($id);
+        if($imovel->user->id !== Auth::user()->id){
+            return redirect('/')->withErrors(['msg'=>'Você só pode alterar os seus próprios imóveis']);
+        }
 
         return view('imoveis.remove', compact('imovel'));
     }
